cleanup admin code
